Updating additional consistent case rules:
  1)  architecture_600
  2)  entity_600
