+++ vsg/rules/__init__.py
@@ -42,7 +42,6 @@ from .blank_lines_between_token_pairs import blank_lines_between_token_pairs
 from .n_spaces_before_and_after_tokens import n_spaces_before_and_after_tokens
 from .spaces_before_and_after_tokens_when_bounded_by_tokens import spaces_before_and_after_tokens_when_bounded_by_tokens
 from .consistent_token_case import consistent_token_case
-from .consistent_case_of_tokens_from_between_tokens_applied_to_region import consistent_case_of_tokens_from_between_tokens_applied_to_region
 from .token_prefix import token_prefix
 from .token_prefix_between_tokens import token_prefix_between_tokens
 from .token_suffix import token_suffix
