+++ vsg/tests/entity/rule_600_test_input.fixed.vhd
@@ -8,3 +8,13 @@ entity FIFO is
     O_DATA : in std_logic_vector(G_WIDTH - 1 downto 0)
   );
 end entity;
+
+entity FIFO2 is
+  generic (
+    g_width : natural := 16
+  );
+  port (
+    I_DATA : in std_logic_vector(g_width - 1 downto 0);
+    O_DATA : in std_logic_vector(g_width - 1 downto 0)
+  );
+end entity;
